[Security Solution] finish moving files to flyout folder + cleanup
