--- x-pack/plugins/security_solution/public/common/components/event_details/event_details.tsx
@@ -1,11 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-export enum EventsViewType {
-  osqueryView = 'osquery-results-view',
-  responseActionsView = 'response-actions-results-view',
-}
