--- x-pack/plugins/security_solution/public/common/components/event_details/insights/translations.ts
@@ -1,16 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { i18n } from '@kbn/i18n';
-
-// TODO: MOVE TO FLYOUT FOLDER - https://github.com/elastic/security-team/issues/7462
-export const SUPPRESSED_ALERTS_COUNT_TECHNICAL_PREVIEW = i18n.translate(
-  'xpack.securitySolution.alertDetails.overview.insights.suppressedAlertsCountTechnicalPreview',
-  {
-    defaultMessage: 'Technical Preview',
-  }
-);
