+++ x-pack/plugins/security_solution/public/common/components/markdown_editor/plugins/insight/index.tsx
@@ -43,7 +43,7 @@ import { useKibana } from '../../../../lib/kibana';
 import { useInsightQuery } from './use_insight_query';
 import { useInsightDataProviders, type Provider } from './use_insight_data_providers';
 import { BasicAlertDataContext } from '../../../../../flyout/document_details/left/components/investigation_guide_view';
-import { InvestigateInTimelineButton } from '../../../event_details/table/investigate_in_timeline_button';
+import { InvestigateInTimelineButton } from '../../../event_details/investigate_in_timeline_button';
 import {
   getTimeRangeSettings,
   parseDateWithDefault,
