+++ x-pack/plugins/security_solution/public/flyout/document_details/right/components/mitre_attack.tsx
@@ -8,9 +8,27 @@
 import { EuiFlexGroup, EuiFlexItem, EuiSpacer, EuiTitle } from '@elastic/eui';
 import type { FC } from 'react';
 import React, { useMemo } from 'react';
-import { MITRE_ATTACK_DETAILS_TEST_ID, MITRE_ATTACK_TITLE_TEST_ID } from './test_ids';
-import { getMitreComponentParts } from '../../../../detections/mitre/get_mitre_threat_component';
+import type { Threats } from '@kbn/securitysolution-io-ts-alerting-types';
+import type { SearchHit } from '../../../../../common/search_strategy';
+import { buildThreatDescription } from '../../../../detection_engine/rule_creation_ui/components/description_step/helpers';
 import { useDocumentDetailsContext } from '../../shared/context';
+import { MITRE_ATTACK_DETAILS_TEST_ID, MITRE_ATTACK_TITLE_TEST_ID } from './test_ids';
+
+/**
+ * Retrieves mitre attack information from the alert information
+ */
+const getMitreComponentParts = (searchHit?: SearchHit) => {
+  const ruleParameters = searchHit?.fields
+    ? searchHit?.fields['kibana.alert.rule.parameters']
+    : null;
+  const threat = ruleParameters ? (ruleParameters[0]?.threat as Threats) : null;
+  return threat && threat.length > 0
+    ? buildThreatDescription({
+        label: threat[0].framework,
+        threat,
+      })
+    : null;
+};
 
 export const MitreAttack: FC = () => {
   const { searchHit } = useDocumentDetailsContext();
