+++ x-pack/plugins/security_solution/public/timelines/components/timeline/tabs/session/use_session_view.tsx
@@ -11,6 +11,9 @@ import styled from 'styled-components';
 import { useDispatch } from 'react-redux';
 import { dataTableSelectors, tableDefaults } from '@kbn/securitysolution-data-table';
 import type { TableId } from '@kbn/securitysolution-data-table';
+import { useExpandableFlyoutApi } from '@kbn/expandable-flyout';
+import { DocumentDetailsRightPanelKey } from '../../../../../flyout/document_details/shared/constants/panel_keys';
+import { useSourcererDataView } from '../../../../../sourcerer/containers';
 import {
   getScopedActions,
   isActiveTimeline,
@@ -20,7 +23,6 @@ import {
 import { useKibana } from '../../../../../common/lib/kibana';
 import * as i18n from './translations';
 import { TimelineTabs } from '../../../../../../common/types/timeline';
-import { useDetailPanel } from '../../../side_panel/hooks/use_detail_panel';
 import { SourcererScopeName } from '../../../../../sourcerer/store/model';
 import { isFullScreen } from '../../body/column_headers';
 import { SCROLLING_DISABLED_CLASS_NAME } from '../../../../../../common/constants';
@@ -242,7 +244,7 @@ export const useSessionViewNavigation = ({ scopeId }: { scopeId: string }) => {
 };
 
 export const useSessionView = ({ scopeId, height }: { scopeId: string; height?: number }) => {
-  const { sessionView } = useKibana().services;
+  const { sessionView, telemetry } = useKibana().services;
   const getScope = useMemo(() => {
     if (isTimelineScope(scopeId)) {
       return timelineSelectors.getTimelineByIdSelector();
@@ -280,10 +282,30 @@ export const useSessionView = ({ scopeId, height }: { scopeId: string; height?: 
       return SourcererScopeName.default;
     }
   }, [scopeId]);
-  const { openEventDetailsPanel } = useDetailPanel({
-    sourcererScope,
-    scopeId,
-  });
+
+  const { selectedPatterns } = useSourcererDataView(sourcererScope);
+  const eventDetailsIndex = useMemo(() => selectedPatterns.join(','), [selectedPatterns]);
+
+  const { openFlyout } = useExpandableFlyoutApi();
+  const openAlertDetailsFlyout = useCallback(
+    (eventId?: string, onClose?: () => void) => {
+      openFlyout({
+        right: {
+          id: DocumentDetailsRightPanelKey,
+          params: {
+            id: eventId,
+            indexName: eventDetailsIndex,
+            scopeId,
+          },
+        },
+      });
+      telemetry.reportDetailsFlyoutOpened({
+        location: scopeId,
+        panel: 'right',
+      });
+    },
+    [openFlyout, eventDetailsIndex, scopeId, telemetry]
+  );
 
   const sessionViewComponent = useMemo(() => {
     const sessionViewSearchBarHeight = 118;
@@ -291,7 +313,7 @@ export const useSessionView = ({ scopeId, height }: { scopeId: string; height?: 
     return sessionViewConfig !== null
       ? sessionView.getSessionView({
           ...sessionViewConfig,
-          loadAlertDetails: openEventDetailsPanel,
+          loadAlertDetails: openAlertDetailsFlyout,
           isFullScreen: fullScreen,
           height: heightMinusSearchBar,
           canReadPolicyManagement,
@@ -301,13 +323,13 @@ export const useSessionView = ({ scopeId, height }: { scopeId: string; height?: 
     height,
     sessionViewConfig,
     sessionView,
-    openEventDetailsPanel,
+    openAlertDetailsFlyout,
     fullScreen,
     canReadPolicyManagement,
   ]);
 
   return {
-    openEventDetailsPanel,
+    openEventDetailsPanel: openAlertDetailsFlyout,
     SessionView: sessionViewComponent,
   };
 };
